Move json error handler to new routing-utils pkg
