+++ packages/identity-x/package.json
@@ -6,6 +6,7 @@
     "@base-cms/marko-web": "^0.9.36",
     "@base-cms/object-path": "^0.9.0",
     "@base-cms/utils": "^0.9.0",
+    "@endeavorb2b/base-website-routing-utils": "^0.11.2",
     "apollo-cache-inmemory": "^1.6.0",
     "apollo-client": "^2.6.0",
     "apollo-link-http": "^1.5.14",
