+++ packages/routing-utils/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@endeavorb2b/base-website-routing-utils",
+  "version": "0.11.2",
+  "private": true,
+  "devDependencies": {
+    "eslint": "^5.15.1",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.16.0"
+  }
+}
